+++ src/Pages/Home.tsx
@@ -1,0 +1,19 @@
+import { ContentArea } from "../components/contentArea/ContentArea";
+import { NavMenu } from "../components/navigation/NavMenu";
+import { Box } from "@mui/material";
+import { appWrapper } from "../styles/styles";
+import { useState } from "react";
+
+export const Home = () => {
+  const [openSide, setOpenSide] = useState(false);
+
+  const handleSideToogle = () => {
+    setOpenSide(!openSide);
+  };
+  return (
+    <Box sx={appWrapper}>
+      <NavMenu handleSideToggle={handleSideToogle} />
+      <ContentArea open={openSide} />
+    </Box>
+  );
+};
